+++ splunk-otel-android/src/main/java/io/opentelemetry/rum/internal/OpenTelemetryRum.java
@@ -16,6 +16,8 @@
 
 package io.opentelemetry.rum.internal;
 
+import android.app.Application;
+
 import io.opentelemetry.api.OpenTelemetry;
 
 /**
@@ -26,9 +28,13 @@ import io.opentelemetry.api.OpenTelemetry;
  */
 public interface OpenTelemetryRum {
 
-    /** Returns a new {@link OpenTelemetryRumBuilder} for {@link OpenTelemetryRum}. */
-    static OpenTelemetryRumBuilder builder() {
-        return new OpenTelemetryRumBuilder();
+    /**
+     * Returns a new {@link OpenTelemetryRumBuilder} for {@link OpenTelemetryRum}.
+     *
+     * @param application The {@link Application} that is being instrumented.
+     */
+    static OpenTelemetryRumBuilder builder(Application application) {
+        return new OpenTelemetryRumBuilder(application);
     }
 
     /** Returns a no-op implementation of {@link OpenTelemetryRum}. */